Don't edit local versions of systemd files.
The current command edits the systemd files inline and then copies them
over.  The problem with this is it leaves the files in an edited state
and they could easily be checked in with other changes.

Instead, edit the files and then write them directly to their
destination instead of saving back to the original file first.